feat: use cdn to serve static js/css better
worth noting that GH Pages only caches pages for 10 minutes, so if we presume that's the time it takes to read an article, the end-user will have to redownload every single resource for each new article. But the cdn servers set the cache header to cache for a year, so yee.