fix(workflow): use site branch for repo-wide artifact
I just realised... there's this one annoyance with artifacts-- they're scoped to the same workflows. So artifacts are more convenient to share between jobs, but between workflows, we either use actions cache, a repo branch, or suffer under the REST API. I should probably document this somewhere so that I don't make the same mistake reimplementing stuff in the future.