perf(build): (fail) attempt using domino instead of cheerio for `wordcountFocused`. see desc.
After dealing with the imgThumbnail build bottleneck, the next highest build bottleneck is in `wordcountFocused`, which deals with counting words in an article. Optimising it by changing the library to domino did not work out well. In particular, the mean average rose by 2ms per operation (from ~12ms to ~14ms) whereas the median average seems to fluctuate. These are rather bad metrics though. What seems to be actually happening is that short content performs well, while large content suffers due to slower removal of nodes (e.g. code, captions, footnotes, appendix) because removal is iterative rather than calling a `.remove()` on an array of nodes, which... idk... is perhaps faster in the cheerio implementation.